% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{densityMixture}
\alias{densityMixture}
\title{C++ function for building a density mixture, given mixture weights and functions.}
\usage{
densityMixture(weights, densities)
}
\description{
C++ function for building a density mixture, given mixture weights and functions.
}
\keyword{internal}
