% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_util.R
\name{fast_ft}
\alias{fast_ft}
\title{FFT: Compute F_n X_n with the real-valued Fourier matrix F_n}
\usage{
fast_ft(x)
}
\description{
FFT: Compute F_n X_n with the real-valued Fourier matrix F_n
}
\keyword{internal}
