% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{lpost}
\alias{lpost}
\title{Unnormalised log posterior}
\usage{
lpost(omega, FZ, k, v, w, u, z, tau, k.theta, MG, G0.alpha, G0.beta, MH,
  H0.alpha, H0.beta, tau.alpha, tau.beta, pdgrm, degree, recompute, db.list)
}
\description{
Unnormalised log posterior
}
\keyword{internal}
