% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{lprior}
\alias{lprior}
\title{Unnormalised log joint prior}
\usage{
lprior(k, v, w, u, z, tau, k.theta, MG, G0.alpha, G0.beta, MH, H0.alpha,
  H0.beta, tau.alpha, tau.beta)
}
\description{
Unnormalised log joint prior
}
\keyword{internal}
