% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureWeight}
\alias{mixtureWeight}
\title{C++ function for computing mixture weights of Bernstein-Mixtures given the probabilities p, values w, and degree k.}
\usage{
mixtureWeight(p, w, k)
}
\description{
C++ function for computing mixture weights of Bernstein-Mixtures given the probabilities p, values w, and degree k.
}
\keyword{internal}
