% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pFromV}
\alias{pFromV}
\title{C++ function for generating  p from v in Stick Breaking DP representation}
\usage{
pFromV(v)
}
\description{
C++ function for generating  p from v in Stick Breaking DP representation
}
\keyword{internal}
