% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_util.R
\name{psd_arma}
\alias{psd_arma}
\title{Analytical spectral density for mean-centred ARMA(p,q) model}
\usage{
psd_arma(freq, ar, ma, sigma2)
}
\description{
Analytical spectral density for mean-centred ARMA(p,q) model
}
\keyword{internal}
