% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_gibbs_core.R
\name{qpsd}
\alias{qpsd}
\title{Compute unnormalised PSD using random mixture of B-splines}
\usage{
qpsd(omega, k, v, w, u, z, degree, recompute, db.list)
}
\description{
Compute unnormalised PSD using random mixture of B-splines
}
\keyword{internal}
