% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unrollPsd}
\alias{unrollPsd}
\title{C++ help function to redundantly roll out a PSD to length n}
\usage{
unrollPsd(qPsd, n)
}
\description{
C++ help function to redundantly roll out a PSD to length n
}
\keyword{internal}
